requirements.txt
browserid
pytest==2.9.1
pytest-selenium
pytest-xdist==1.14
requests==2.9.1
